Move LinesWithClickableInterior to Example for now
