+++ docs/src/4.6.Lines.mdx
@@ -6,12 +6,9 @@ import { LinesDemo, LinesWireframe, LinesHitmap } from "./jsx/allLiveEditors";
 
 ## Props
 
-| Name                      | Type      | Default                      | Description                                                                                                                                     |
-| ------------------------- | --------- | ---------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------- |
-| `children`                | `Line[]`  | `[]`                         | array of Line markers to render                                                                                                                 |
-| `enableClickableInterior` | `boolean` |                              | when enabled, a polygon will be drawn using the line `points`, and the user will get the original line object after clicking inside the polygon |
-| `fillColor`               | `Color`   | `{ r: 0, g: 0, b: 0, a: 0 }` | the fill color for the polygon                                                                                                                  |
-| `showBorder`              | `boolean` |                              | draw the lines around the polygon                                                                                                               |
+| Name       | Type     | Default | Description                     |
+| ---------- | -------- | ------- | ------------------------------- |
+| `children` | `Line[]` | `[]`    | array of Line markers to render |
 
 ### Line
