+++ packages/regl-worldview/src/commands/Lines.js
@@ -6,16 +6,13 @@
 //  found in the LICENSE file in the root directory of this source tree.
 //  You may not use this file except in compliance with the License.
 
-import React from "react";
-
-import type { Line, Color } from "../types";
+import type { Line } from "../types";
 import { defaultBlend, withPose, toRGBA, shouldConvert, pointToVec3 } from "../utils/commandUtils";
 import {
   getHitmapPropsForInstancedCommands as getHitmapProps,
   getObjectForInstancedCommands as getObjectFromHitmapId,
 } from "../utils/hitmapDefaults";
-import { makeCommand, type Props } from "./Command";
-import FilledPolygons from "./FilledPolygons";
+import { makeCommand } from "./Command";
 
 /*
 Triangle-based line drawing.
@@ -440,48 +437,4 @@ const Lines = makeCommand<Line>("Lines", lines, {
   getObjectFromHitmapId,
 });
 
-export type LineProps = Props<Line> & {
-  enableClickableInterior?: boolean,
-  fillColor: Color, // visually turn lines into polygons using custom fillColor
-  showBorder: boolean,
-};
-
-function LinesWithClickableInterior({
-  children,
-  enableClickableInterior,
-  fillColor,
-  onClick,
-  showBorder,
-  ...rest
-}: LineProps) {
-  if (enableClickableInterior) {
-    return (
-      <>
-        {showBorder && <Lines {...rest}>{children}</Lines>}
-        <FilledPolygons
-          onClick={(ev, { object, ...rest }) => {
-            onClick(ev, { ...rest, object: object.lineObject, objectId: object.lineObject.id });
-          }}>
-          {children.map((item) => ({
-            id: item.id,
-            points: item.points,
-            lineObject: item,
-            color: fillColor,
-          }))}
-        </FilledPolygons>
-      </>
-    );
-  }
-  return (
-    <Lines {...rest} onClick={onClick}>
-      {children}
-    </Lines>
-  );
-}
-
-LinesWithClickableInterior.defaultProps = {
-  fillColor: { r: 0, g: 0, b: 0, a: 0 },
-  showBorder: false,
-};
-
-export default LinesWithClickableInterior;
+export default Lines;
